new main functionalities + broken streaming
What this commit doesn't do: work.
What this commit does do: adds the logic to decipher whether an email should be fired or not and the basic framework to do so. The folder listener utilizes a scriptable stream, and streams the incoming message, but breaks as soon as we try to read it (it doesn't do this with any other messages.) I've been working on this bug for over a week, and I'm out of ideas and resources. I think the next step is to branch to checking for encryption upon clicking an unread message as opposed to receiving an unread message, as the header seems to stream fine for existing messages. This will require a rework of most of the existing listener code, but might be the only option at this point.