CryptCheck detects incoming unencrypted messages
After a week and a half, checking for encryption in the incoming message finally works. It's hacky and objectively bad, but I'm blaming this on the XPCOM utilities for now. The folder and message key of new messages in the inbox are added to a stack, which is checked every 10 seconds (as of now); if the stack is not empty, the cryptCheck function is called until it is. CryptCheck pops from the stack, finds the messageKey in the folder, and checks to see if the message breaks protocol. I wanted to stream the message to check for encryption in the folder listener but it timed out regardless of implementation (my suspicion is that the message wasn't fully loaded into the folder yet.) This implementation, although roundabout and potentially unscalable, enables work on the next part of the extension: composing and sending the nastygram.